`Selectable` mixin and `SelectableEventedList` (#2439)
* selectable mixin and selectable evented list

* add group

* typo
