+++ napari/utils/events/containers/__init__.py
@@ -1,6 +1,7 @@
 from ._evented_list import EventedList
 from ._nested_list import NestableEventedList
-from ._selection import Selection
+from ._selectable_list import SelectableEventedList
+from ._selection import Selectable, Selection
 from ._set import EventedSet
 from ._typed import TypedMutableSequence
 
@@ -9,5 +10,7 @@ __all__ = [
     'EventedSet',
     'NestableEventedList',
     'Selection',
+    'SelectableEventedList',
+    'Selectable',
     'TypedMutableSequence',
 ]
